tts-reader/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
websockets
PyMuPDF
pdfminer.six
spacy
numpy
scikit-learn
gTTS
pydub